+++ services/workflows-service/src/collection-flow/controllers/collection-flow.business.controller.ts
@@ -15,7 +15,6 @@ export class CollectionFlowBusinessController {
     const { countryCode, state, vendor, registrationNumber } = query;
 
     const jurisdictionCode = this.buildJurisdictionCode(countryCode, state);
-    console.log('CODE', jurisdictionCode, typeof state, query);
 
     return this.businessService.fetchCompanyInformation({
       registrationNumber,