+++ services/workflows-service/src/workflow/dtos/document-decision-query.input.ts
@@ -1,0 +1,9 @@
+import { oneOf } from '@/common/decorators/one-of.decorator';
+import { IsOptional, IsString } from 'class-validator';
+
+export class DocumentDecisionUpdateQueryInput {
+  @IsString()
+  @IsOptional()
+  @oneOf(['base', 'director'])
+  contextUpdateMethod!: 'base' | 'director';
+}