+++ apps/kyb-app/CHANGELOG.md
@@ -61,6 +61,7 @@
 ### Patch Changes
 
 - Updated dependencies [8c888cec]
+- Updated dependencies
   - @ballerine/workflow-browser-sdk@0.5.22
   - @ballerine/blocks@0.1.21
   - @ballerine/common@0.7.23
@@ -84,6 +85,8 @@
 
 ### Patch Changes
 
+- @ballerine/blocks@0.1.20
+- @ballerine/workflow-browser-sdk@0.5.20
 - Change document schemas
 - Updated dependencies
   - @ballerine/blocks@0.1.20