+++ packages/ui/CHANGELOG.md
@@ -17,6 +17,7 @@
 ### Patch Changes
 
 - 8c888cec: Updated tsconfig and types
+- Fixed missing disabled state on inputs
 
 ## 0.3.14